move custom code to children resources
# Conflicts:
#	front/lib/resources/group_space_resource.ts
