+++ front/lib/resources/group_resource.ts
@@ -48,6 +48,20 @@ import type { LightWorkspaceType, UserType } from "@app/types/user";
 export const ADMIN_GROUP_NAME = "dust-admins";
 export const BUILDER_GROUP_NAME = "dust-builders";
 
+/**
+ * ┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┓
+ * ┃                                                                         ┃
+ * ┃  IMPORTANT: GroupResource DOES NOT and SHOULD NOT have permissions      ┃
+ * ┃  management of its own.                                                 ┃
+ * ┃                                                                         ┃
+ * ┃  Groups are designed to be used within the context of other resources   ┃
+ * ┃  (e.g., SpaceResource, AgentConfigurationResource). The permissions     ┃
+ * ┃  should be managed at the junction with parent resource level,          ┃
+ * ┃  not at the group level.                                                ┃
+ * ┃                                                                         ┃
+ * ┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┛
+ */
+
 // Attributes are marked as read-only to reflect the stateless nature of our Resource.
 // This design will be moved up to BaseResource once we transition away from Sequelize.
 // eslint-disable-next-line @typescript-eslint/no-unsafe-declaration-merging
@@ -1265,7 +1279,7 @@ export class GroupResource extends BaseResource<GroupModel> {
         this.kind === "agent_editors" ||
         this.kind === "skill_editors" ||
         (allowProvisionnedGroups && this.kind === "provisioned"),
-      `You can't add members to ${this.kind} groups.`
+      `You can't remove members to ${this.kind} groups.`
     );
     const owner = auth.getNonNullableWorkspace();
     if (users.length === 0) {
